requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.3.7.1
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.3
Pillow==2.6.1
PyJWT==1.7.1
RPi.GPIO==0.6.3
RTIMULib==7.2.1
Werkzeug==0.9.6
argparse==1.2.1
blinker==1.3
chardet==2.3.0
colorama==0.3.2
dominate==2.3.5
gpiozero==1.4.0
html5lib==0.999
itsdangerous==0.24
lxkeymap==0.1
lxml==3.4.0
mcpi==0.1.1
mercurial==3.1.2
ndg-httpsclient==0.3.2
numpy==1.8.2
pbr==3.1.1
picamera==1.13
picraft==1.0
pifacecommon==4.2.1
pifacedigitalio==3.1.0
pigpio==1.35
pyOpenSSL==0.13.1
pyasn1==0.1.7
pygame==1.9.2a0
pygobject==3.14.0
pyimgur==0.6.0
pyinotify==0.9.4
pyprof2calltree==1.4.4
pyserial==2.6
python-apt==0.9.3.12
pytz==2018.7
requests==2.4.3
sense-emu==1.0
sense-hat==2.2.0
six==1.10.0
smbus==1.1
spidev==3.0
stevedore==1.23.0
twilio==6.21.0
twython==3.1.2
urllib3==1.9.1
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
visitor==0.1.3
wheel==0.24.0
wsgiref==0.1.2